(sitka) [RT25288] line item number and PO note
This commit removes the condition which limits the index being used to
create the LIN+# value.  These values should start at 1 and increase by
1 per line item in the PO according to the standard here:
http://www.editeur.org/files/EDIfact%20eancom%20pdfs/EDIfact%20library%20supply/L4orders%20%28library%20supply%29.pdf

As well, this adds code to include a PO note, which is separate from
lineitem notes.  The standard says not to send free text notes without a
specific 4441 value if possible, but EG offers the option of a PO Note,
so this is not possible in our case.